move eventloops wrappers to eventloop file
