+++ eventloop.c
@@ -42,6 +42,7 @@
 #include <sys/time.h>
 
 #include "eventloop.h"
+#include "indidevapi.h"
 
 /* info about one registered callback.
  * the malloced array cback is never shrunk, entries are reused. new id's are
@@ -445,6 +446,49 @@ static void oneLoop()
 static void deferTO(void *p)
 {
     *(int *)p = 1;
+}
+
+
+/* "INDI" wrappers to the more generic eventloop facility. */
+
+int IEAddCallback(int readfiledes, IE_CBF *fp, void *p)
+{
+    return (addCallback(readfiledes, (CBF *)fp, p));
+}
+
+void IERmCallback(int callbackid)
+{
+    rmCallback(callbackid);
+}
+
+int IEAddTimer(int millisecs, IE_TCF *fp, void *p)
+{
+    return (addTimer(millisecs, (TCF *)fp, p));
+}
+
+void IERmTimer(int timerid)
+{
+    rmTimer(timerid);
+}
+
+int IEAddWorkProc(IE_WPF *fp, void *p)
+{
+    return (addWorkProc((WPF *)fp, p));
+}
+
+void IERmWorkProc(int workprocid)
+{
+    rmWorkProc(workprocid);
+}
+
+int IEDeferLoop(int maxms, int *flagp)
+{
+    return (deferLoop(maxms, flagp));
+}
+
+int IEDeferLoop0(int maxms, int *flagp)
+{
+    return (deferLoop0(maxms, flagp));
 }
 
 #if defined(MAIN_TEST)
